+++ ConfluencePS/Private/ConvertTo-ServerInfo.ps1
@@ -1,0 +1,35 @@
+﻿function ConvertTo-ServerInfo {
+    [CmdletBinding()]
+    [OutputType([ConfluencePS.ServerInfo])]
+    param (
+        [Parameter(ValueFromPipeline = $true)]
+        [PSObject[]]$InputObject
+    )
+
+    PROCESS {
+        foreach ($object in $InputObject) {
+            Write-Verbose "[$($MyInvocation.MyCommand.Name)] Converting Object to ServerInfo"
+
+            $deploymentType = if ($object.deploymentType) { $object.deploymentType } elseif ($object.cloudId) { 'Cloud' } else { 'DataCenter' }
+            $hash = @{
+                DeploymentType  = $deploymentType
+                CloudId         = $object.cloudId
+                CommitHash      = $object.commitHash
+                Edition         = $object.edition
+                SiteTitle       = $object.siteTitle
+                DefaultLocale   = $object.defaultLocale
+                DefaultTimeZone = $object.defaultTimeZone
+                MicrosPerimeter = $object.microsPerimeter
+                Version         = $object.version
+            }
+
+            if ($object.baseUrl) { $hash.BaseUrl = [Uri]$object.baseUrl }
+            if ($object.fallbackBaseUrl) { $hash.FallbackBaseUrl = [Uri]$object.fallbackBaseUrl }
+            if ($object.buildNumber) { $hash.BuildNumber = [Int32]$object.buildNumber }
+            if ($object.buildDate) { $hash.BuildDate = [DateTime]$object.buildDate }
+            if ($object.serverTime) { $hash.ServerTime = [DateTime]$object.serverTime }
+
+            [ConfluencePS.ServerInfo]$hash
+        }
+    }
+}